requirements.txt
bleak>=0.21.0
customtkinter>=5.2.0
Pillow>=10.0.0
pyinstaller>=6.0.0
